refactor(version manager): move version manager to its own file
